moved util.ts to util folder
